Change diffs working.
All the pieces are present, but it is not quite useful.  The coverage
formatter will fire an exception only if the uncovered source lines
and the changed lines match exactly.  Not really useful yet.

The equality check between two CodeChange objects needs to return true
if there is a non-empty intersection between the two source listings.